This commit resolves #9 and SCPR/KPCC-iPad#11.
To keep parity with the currently shipped iPad version, ads will now disappear if the user rotates the device while they display. This keeps from needing to fetch a new creative at a different size.
Also added in this commit is the implementation of Apple's new standard of requesting notification settings permission in iOS 8.